requirements.txt
pandas
numpy
gymnasium
sphinx
pytest
git+https://github.com/carlosluis/stable-baselines3@fix_tests
PyYAML
wandb